Moves all build logic into separate functions.
