+++ build-logic/convention/src/main/kotlin/com/revenuecat/purchases/kmp/buildlogic/ktx/Project.kt
@@ -1,0 +1,9 @@
+package com.revenuecat.purchases.kmp.buildlogic.ktx
+
+import org.gradle.api.Project
+import org.gradle.api.artifacts.VersionCatalog
+import org.gradle.api.artifacts.VersionCatalogsExtension
+import org.gradle.kotlin.dsl.getByType
+
+internal val Project.versionCatalog: VersionCatalog
+    get() = extensions.getByType<VersionCatalogsExtension>().named("libs")
